Pipe: a new kind of `Stage`
Add `Pipe`, a `Stage` that itself runs multiple other stages with
their stdouts and stdins piped to one another. This is a useful
concept, because it allows pipelines to be created in modular form
more easily.

Also, now that we have this type, we can is it in the implementation
of `Pipeline`. Specifically, `Pipeline` now embeds a single `Pipe`,
which takes care of managing multiple stages and piping them together.
This means that `Pipeline` only needs to know how to run a single
stage.

More changes are still to come…
